fix:TR-4934 whole changes kept at service level, based on Kiryl review
